Add permission CRUD with modal form

+++ app/Livewire/Buildings/BuildingIndex.php
@@ -12,11 +12,19 @@ class BuildingIndex extends Component
 
     public function mount()
     {
+        // Permission: building.list
+        if (!auth()->user()?->can('building.list')) {
+            abort(403, 'You do not have permission to view buildings.');
+        }
         $this->buildings = Building::all();
     }
 
     public function delete($id)
     {
+        // Permission: building.delete
+        if (!auth()->user()?->can('building.delete')) {
+            abort(403, 'You do not have permission to delete buildings.');
+        }
         Building::findOrFail($id)->delete();
         $this->buildings = Building::all();
         $this->dispatch('flash-success', message: __('Building deleted successfully.'));
@@ -25,6 +33,10 @@ class BuildingIndex extends Component
     #[On('building-created')]
     public function refreshBuildings($url)
     {
+        // Permission: building.list
+        if (!auth()->user()?->can('building.list')) {
+            abort(403, 'You do not have permission to view buildings.');
+        }
         $this->buildings = Building::all();
         $this->dispatch('navigate', $url);
     }
@@ -37,6 +49,10 @@ class BuildingIndex extends Component
 
     public function render()
     {
+        // Permission: building.list
+        if (!auth()->user()?->can('building.list')) {
+            abort(403, 'You do not have permission to view buildings.');
+        }
         return view('livewire.buildings.building-index');
     }
 }
